pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.60.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.2.5",
    "serpapi>=0.1.5",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
]
